Extract AccordionBlock into its own component file
Move the accordionBlock renderer out of docAndBlogComponents into
templateComponents/accordionBlock.tsx. The component map passes its
components prop in to render nested rich-text, avoiding a circular import.
